test: removed extraneous dependencies because of profiling tests
Solution: moved the profiling test to a separate module.

This sets a temporary trade-off, as a lot of test code (all test cases)
have been hastily duplicated. We'll see in a forthcoming PR how to
refactor test cases so they may be reusable by multiple modules.

NOTE: it doesn't hurt if the new sub module requires a higher version of
go.

Signed-off-by: Frederic BIDON <fredbi@yahoo.com>